+++ messaging/rabbit/consumer_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 //nolint:errcheck // Tests: errores de Close()/Wait() en cleanup se ignoran intencionalmente
 package rabbit