+++ testing/containers/helpers_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 //nolint:errcheck // Tests: errores de Close() en cleanup se ignoran intencionalmente
 package containers